SeparateCodeAndWriteInterface
